[feature] 환자 수정 추가

+++ src/main/kotlin/com/example/hdjunctiontest/domain/entity/hospital/Hospitals.kt
@@ -38,7 +38,7 @@ class Hospitals(
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: BigInteger,
+    val id: Long,
 ) : BaseEntity() {
 
     fun delete() {
